requirements.txt
requests==2.32.5
httpx==0.27.2
beautifulsoup4==4.14.3
playwright==1.58.0
python-dateutil==2.9.0.post0
lxml==6.0.2
feedparser==6.0.12
openai==2.28.0
python-dotenv==1.2.2
supabase==2.15.1
fastapi==0.115.6
uvicorn==0.34.0
